+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Added
+* Check extent when adding layer to tile
 
 ## [0.1.0] - 2010-01-10
 * Initial version